Added new activation functions: * Parametrized ReLu, by adding a leakage value we can go from normal ReLu (0) to Leaky (0,001) or any other value we consider necesary. That's the whole point of it being parametrized.
* Softmax, added two softmax functions. One that is unstable for problem
explanation. One that is balanced to show how it fixes the issue.
Both still lack the derivative since  the first is not intended to be
used so is not gonna get it and I will add it to the second later.

Some other fixes and changed to the code were made.
More comments and other stuff.